Instrumentation refactoring
Use bool instead of int where it should be.
Tweak naming.
Removed unused members

+++ unit_tests/os_interface/linux/mock_performance_counters_linux.cpp
@@ -66,7 +66,7 @@ int dlcloseMock(void *handle) throw() {
     }
     return 0;
 }
-uint64_t setPlatformInfo(uint32_t productId, void *pSkuTable) {
+uint32_t setPlatformInfo(uint32_t productId, void *pSkuTable) {
     PerfCounterFlagsLinux::setPlatformInfoFuncCalled++;
     return 0;
 }
