+++ unit_tests/os_interface/linux/mock_performance_counters_linux.h
@@ -32,7 +32,7 @@ void *dlsymMockPassing(void *handle, const char *symbol) throw();
 void *dlsymMockFailing(void *handle, const char *symbol) throw();
 int getTimeFuncPassing(clockid_t clk_id, struct timespec *tp) throw();
 int dlcloseMock(void *handle) throw();
-uint64_t setPlatformInfo(uint32_t productId, void *pSkuTable);
+uint32_t setPlatformInfo(uint32_t productId, void *pSkuTable);
 int perfmonLoadConfigMock(int fd, drm_intel_context *ctx, uint32_t *oa_cfg_id, uint32_t *gp_cfg_id);
 
 class PerfCounterFlagsLinux : public PerfCounterFlags {
